+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 <!-- The next version number should be the version already set in pubspec.yaml -->
 
+## [1.11.2] - 2025-09-18
+### Changed
+- Acknowledge Comment innerContext field and pass it when create/delete/load comments [#5356](https://github.com/rokwire/illinois-app/issues/5356).
+
+## [1.11.1] - 2025-09-17
+### Fixed
+- Survey result rule evaluation issues [#637](https://github.com/rokwire/app-flutter-plugin/issues/637).
+### Added
+- Acknowledge Reaction innerContext field [#5238](https://github.com/rokwire/illinois-app/issues/5238). 
+
 ## [1.11.0] - 2025-08-05
 ### Changed
 - Android: upgrade compileSdkVersion and targetSdkVersion to 36 [#626](https://github.com/rokwire/app-flutter-plugin/issues/626).